doc_processing_worker/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
redis==5.0.3
rq==2.6.0
langchain==0.2.7
langchain-community
langchain-google-genai==1.0.10
chromadb==0.4.24
pypdf==4.1.0
Pillow
pytesseract==0.3.10
psycopg2-binary==2.9.9
pysqlite3-binary